% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/licorice.R
\name{get_middle_position}
\alias{get_middle_position}
\title{find the coordinated to center the responses in a center plot}
\usage{
get_middle_position(middle_pos, levels, start_values, end_values)
}
\arguments{
\item{middle_pos}{a number corresponding to the location at which to center the plot}

\item{levels}{the used response levels}

\item{start_values}{the y start values}

\item{end_values}{the y end values}
}
\description{
find the coordinated to center the responses in a center plot
}

