% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/licorice.R
\name{licorice}
\alias{licorice}
\title{plot the likert scale results using centered bar plot}
\usage{
licorice(data, answers_order = levels(as.factor(data$response)),
  sort = FALSE, sort_reversed = FALSE, middle_pos = 2.5,
  type = c("center", "fill", "count"), strwrap_width = 25)
}
\arguments{
\item{data}{the dataset}

\item{sort}{whether to sort results, defaults to TRUE}

\item{middle_pos}{a number corresponding to the location at which to center the plot}

\item{type}{the type of graph to create (center, fill, count)}

\item{strwrap_width}{at what position in the question to place newline characters}

\item{short_reversed}{should sort be reversed?}
}
\description{
plot the likert scale results using centered bar plot
}

