% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{licorice_center_plot}
\alias{licorice_center_plot}
\title{Centered licorice plot}
\usage{
licorice_center_plot(data_transformed)
}
\arguments{
\item{data_transformed}{the transformed dataset}
}
\description{
Centered licorice plot
}

