% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{licorice_count_plot}
\alias{licorice_count_plot}
\title{Count licorice plot}
\usage{
licorice_count_plot(data_transformed)
}
\arguments{
\item{data_transformed}{the transformed dataset}
}
\description{
Count licorice plot
}

