% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{licorice_fill_plot}
\alias{licorice_fill_plot}
\title{Filled licorice plot}
\usage{
licorice_fill_plot(data_transformed)
}
\arguments{
\item{data_transformed}{the transformed dataset}
}
\description{
Filled licorice plot
}

