% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/licorice.R
\name{transform_data}
\alias{transform_data}
\title{Transforms data to be used by licorice function}
\usage{
transform_data(data, answers_order, sort = TRUE, middle_pos = 2.5,
  strwrap_width = 25)
}
\arguments{
\item{data}{dataset}

\item{answers_order}{character vector with answers in order}

\item{sort}{sort data?}

\item{middle_pos}{middle position of the factor var}
}
\description{
Transforms data to be used by licorice function
}

